feat: Add android and ios directories to .gitignore
Ignore the android and ios directories as well as the .yarn directory in the .gitignore file. This ensures that these directories and files are not tracked by Git, preventing them from being included in the repository. This is necessary because these directories and files are typically generated or specific to individual developers and should not be shared in the repository.